Adapt highlightAreaPlot to new atlas and add options
The new atlas has for instance a dLGN shell and dLGN core. That
makes 2 match for the area and crashes the area plot. I could
make the match to be exact but then shell and core would be different
colors. I prefer to be more lax and allow partial match of multiple
area. As a side effect giving "layer 2" for instance will highlight
all the areas with layer 2. That will probably crash if one area
matches multiple string given in the came call.

Add options, mostly an `alpha` argument to make the areas slighlty
transparent